+++ NEWS.md
@@ -1,5 +1,9 @@
 # TARDIS 2.0
 
+## Changes in 2.2.0b
+- Fixed issue: detecting only the highest peak
+- Ignore local maxima that is lower than 10% of the maximum peak intensity, not 50% (previous)
+
 ## Changes in 2.1.0b
 - Improved algorithm efficiency and parallel computing code
 - Achieved lower runtime in comparison to the original code by Pablo